Fix port range check and move Sip003Plugin.
